new foldr and task 3

+++ Lesso_one/triangle_area.rb
@@ -2,5 +2,5 @@ puts "Please enter triangle base"
 triangle_base = gets.chomp
 puts "Please enter triangle hight"
 triangle_hight = gets.chomp
-triangle_area = 0.5 * triangle_hight.to_i * triangle_base.to_i
+triangle_area = 0.5 * triangle_hight.to_f * triangle_base.to_f
 puts "triangle area is #{triangle_area}"
